Move to utils.
