+++ main.py
@@ -1,6 +1,5 @@
 import asyncio
 
-import httpx
 from loguru import logger
 from telegram import (
     BotCommand,
@@ -18,7 +17,7 @@ from telegram.ext import (
     filters,
 )
 
-from hiroshi.config import application_settings, telegram_settings
+from hiroshi.config import telegram_settings
 from hiroshi.services.bot import (
     handle_available_providers_options,
     handle_prompt,
@@ -32,7 +31,7 @@ from hiroshi.utils import (
     get_telegram_chat,
     get_telegram_message,
     log_application_settings,
-    user_interacts_with_bot,
+    user_interacts_with_bot, uptime_checker,
 )
 
 
@@ -165,22 +164,6 @@ class HiroshiBot:
         # )
         app.add_error_handler(self.error_handler)
         app.run_polling()
-
-
-async def uptime_checker() -> None:
-    if application_settings.monitoring_is_active and application_settings.monitoring_url:
-        logger.info(f'Uptime Checker started. '
-                    f'MONITORING_FREQUENCY_CALL={application_settings.monitoring_frequency_call} '
-                    f'MONITORING_URL={application_settings.monitoring_url}')
-        while True:
-            async with httpx.AsyncClient() as client:
-                result = await client.get(application_settings.monitoring_url)
-            if result.status_code != 200:
-                logger.error(f'Uptime Checker failed. status_code({result.status_code}) msg: {result.text}')
-            # Converting from minutes to seconds.
-            await asyncio.sleep(application_settings.monitoring_frequency_call * 60)
-    else:
-        logger.info('Uptime Checker disabled. To turn it on set MONITORING_IS_ACTIVE environment variable.')
 
 
 if __name__ == "__main__":
